feat: remove HoverOverlay component
This library no longer provides the React component for rendering the hover tooltip.

BREAKING CHANGE: Removes the HoverOverlay component. You must create your own UI component that renders from the hoverStateUpdates data.
