+++ src/index.ts
@@ -1,3 +1,2 @@
 export * from './hoverifier'
-export * from './HoverOverlay'
 export * from './positions'
